requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
pyyaml>=6.0
lxml>=5.0
